requirements.txt
# DuwatBench - Arabic Calligraphy Benchmark
# Core dependencies

# Core
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
python-Levenshtein>=0.20.0

# Data handling
datasets>=2.14.0

# API clients (for closed-source models)
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.3.0

# Open-source models (optional - install based on what you need)
# Uncomment the models you want to use:

# transformers>=4.35.0
# torch>=2.0.0
# torchvision>=0.15.0
# accelerate>=0.24.0
# sentencepiece>=0.1.99
# einops>=0.7.0
# timm>=0.9.0

# EasyOCR
# easyocr>=1.7.0

# Visualization (optional)
# matplotlib>=3.7.0
# seaborn>=0.12.0
